+++ src/aasm/__init__.py
@@ -3,6 +3,13 @@ from .model import MachineState, ProblemSpec, TaskEnvelope, Proposal, Result, Ca
 from .agents import AASMAgent, FunctionAgent
 from .authority import SingleControllerAuthority, AutonomousAuthority, QuorumAuthority, HierarchicalAuthority
 from .effects import EffectSpec, EffectRecord, EffectStatus, RetryPolicy, EffectExecutionError, EffectUnknownOutcome
-__all__=["AASMEngine","MachineState","ProblemSpec","TaskEnvelope","Proposal","Result","CapabilitySet","AASMAgent","FunctionAgent","SingleControllerAuthority","AutonomousAuthority","QuorumAuthority","HierarchicalAuthority","MemoryStore","SQLiteStore","EffectSpec","EffectRecord","EffectStatus","RetryPolicy","EffectExecutionError","EffectUnknownOutcome"]
-
+from .definitions import MachineDefinition, default_machine_definition
+from .model_check import ModelCheckIssue, ModelCheckReport, check_machine
 from .persistence import MemoryStore, SQLiteStore
+
+__all__=[
+    "AASMEngine","MachineState","ProblemSpec","TaskEnvelope","Proposal","Result","CapabilitySet",
+    "AASMAgent","FunctionAgent","SingleControllerAuthority","AutonomousAuthority","QuorumAuthority","HierarchicalAuthority",
+    "MemoryStore","SQLiteStore","EffectSpec","EffectRecord","EffectStatus","RetryPolicy","EffectExecutionError","EffectUnknownOutcome",
+    "MachineDefinition","default_machine_definition","ModelCheckIssue","ModelCheckReport","check_machine"
+]